+++ example/variables.tf
@@ -1,3 +1,4 @@
 variable "nomad_acl" {
   type = bool
+  description = "nomad_acl"
 }