+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Added
 - Using Vault to generate username and password for postgres with kv #13
+- Added documentation and LICENSE #2
 
 ## [0.0.1]