+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Using Vault to generate username and password for postgres with kv #13
 - Added host volume for postgres and dynamic variables #12
 - Added switch for host volume #12
+- Added 03_run_tests.yml #3
+- Code to support successful execution of nomad mc job and tests when consul_acl_default_policy is deny #22
+- Added documentation and LICENSE #2
 
 ## [0.0.1]