+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added 03_run_tests.yml #3
 - Code to support successful execution of nomad mc job and tests when consul_acl_default_policy is deny #22
 - Added documentation and LICENSE #2
+- Added tests for host volumes #25
 
 ## [0.0.1]